Revert "[learn, learn_detail] 스타일 분리 작업"
