+++ src/components/learnDetail/ContextMenu.tsx
@@ -1,8 +1,42 @@
-import styled, { css } from 'styled-components';
 import { COLOR } from '@src/styles/color';
 import { FONT_STYLES } from '@src/styles/fontStyle';
+import styled, { css } from 'styled-components';
 
-export const StContextMenu = styled.div<{ top: number; left: number }>`
+interface ContextMenuProps {
+  points: {
+    x: number;
+    y: number;
+  };
+  setIsNewMemo: (create: boolean) => void;
+}
+
+function ContextMenu(props: ContextMenuProps) {
+  const { points, setIsNewMemo } = props;
+
+  return (
+    <StContextMenu top={points.y} left={points.x} className="test">
+      <ul>
+        <li>
+          <button
+            type="button"
+            onClick={(e) => {
+              e.stopPropagation();
+              setIsNewMemo(true);
+            }}>
+            메모 추가
+          </button>
+        </li>
+        <li>
+          <button type="button">하이라이트 삭제</button>
+        </li>
+      </ul>
+    </StContextMenu>
+  );
+}
+
+export default ContextMenu;
+
+const StContextMenu = styled.div<{ top: number; left: number }>`
   display: flex;
   flex-direction: column;
   align-items: center;
