--- src/components/learnDetail/ScriptEdit/style.ts
@@ -1,55 +1,0 @@
-import styled from 'styled-components';
-import { COLOR } from '@src/styles/color';
-
-export const StWrapper = styled.div`
-  position: relative;
-  caret-color: transparent;
-  cursor: pointer;
-
-  :focus {
-    outline: none;
-  }
-
-  & > mark {
-    background: linear-gradient(259.3deg, #d8d9ff 0%, #a7c5ff 100%);
-  }
-`;
-
-export const StScriptText = styled.div`
-  position: relative;
-  font-size: 2.6rem;
-  color: ${COLOR.BLACK};
-  cursor: pointer;
-
-  & > span {
-    font-size: 3.2rem;
-    font-weight: 600;
-    color: #4e8aff;
-  }
-
-  & > .left {
-    margin-right: 0.4rem;
-  }
-
-  & > .right {
-    margin-left: 0.4rem;
-  }
-
-  & > mark {
-    background: linear-gradient(259.3deg, #d8d9ff 0%, #a7c5ff 100%);
-
-    & > span {
-      font-size: 3.2rem;
-      font-weight: 600;
-      color: ${COLOR.MAIN_BLUE};
-    }
-
-    & > .left {
-      margin-right: 0.4rem;
-    }
-
-    & > .right {
-      margin-left: 0.4rem;
-    }
-  }
-`;
