--- src/components/learnDetail/VideoDetail/style.ts
@@ -1,32 +1,0 @@
-import styled from 'styled-components';
-import { COLOR } from '@src/styles/color';
-import { FONT_STYLES } from '@src/styles/fontStyle';
-
-export const StVideoDetail = styled.div`
-  & > div {
-    display: flex;
-    color: ${COLOR.GRAY_30};
-    ${FONT_STYLES.M_21_BODY};
-  }
-
-  & > h1 {
-    margin-top: 1.2rem;
-    margin-bottom: 2rem;
-    color: ${COLOR.BLACK};
-    ${FONT_STYLES.SB_32_HEADLINE};
-  }
-`;
-
-export const StTagContainer = styled.div`
-  display: flex;
-  gap: 0.8rem;
-  margin-bottom: 4.8rem;
-
-  & > span {
-    padding: 1rem 1.6rem;
-    border-radius: 2.4rem;
-    color: ${COLOR.WHITE};
-    background-color: ${COLOR.MAIN_BLUE};
-    ${FONT_STYLES.M_18_CAPTION};
-  }
-`;
